+++ tzrec/modules/sid_generation/__init__.py
@@ -1,0 +1,50 @@
+# Copyright (c) 2024, Alibaba Group;
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#    http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from tzrec.modules.sid_generation.clip_loss import (
+    CLIPLoss,
+    GatherLayer,
+)
+from tzrec.modules.sid_generation.kmeans import (
+    MiniBatchKMeans,
+)
+from tzrec.modules.sid_generation.residual_kmeans import (
+    ResidualKMeans,
+    RQKMeans,
+)
+from tzrec.modules.sid_generation.residual_quantized import (
+    ResidualQuantized,
+)
+from tzrec.modules.sid_generation.rqvae import (
+    RQVAE,
+)
+from tzrec.modules.sid_generation.types import (
+    QuantizeForwardMode,
+    QuantizeOutput,
+    ResidualQuantizedOutput,
+)
+from tzrec.modules.sid_generation.vector_quantize import (
+    VectorQuantize,
+)
+
+__all__ = [
+    "QuantizeForwardMode",
+    "QuantizeOutput",
+    "ResidualQuantizedOutput",
+    "VectorQuantize",
+    "GatherLayer",
+    "CLIPLoss",
+    "ResidualQuantized",
+    "RQVAE",
+    "MiniBatchKMeans",
+    "ResidualKMeans",
+    "RQKMeans",
+]
